+++ src/com/gerardng/networkchat/server/Server.java
@@ -3,9 +3,11 @@ package com.gerardng.networkchat.server;
 import java.io.IOException;
 import java.net.DatagramPacket;
 import java.net.DatagramSocket;
+import java.net.InetAddress;
 import java.net.SocketException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.UUID;
 
 // Thread
 public class Server implements Runnable {
@@ -16,6 +18,7 @@ public class Server implements Runnable {
 	private DatagramSocket datagramSocket;
 	private Thread runThread, manageThread, receiveThread, sendThread;
 	private boolean running = false;
+	private Thread send;
 	
 	public Server(int port) {
 		this.port = port;
@@ -78,11 +81,36 @@ public class Server implements Runnable {
 	private void process(DatagramPacket datagramPacket) {
 		String message = new String(datagramPacket.getData());
 		if(message.startsWith("/c/")) {
-			clients.add(new ServerClient("alias", datagramPacket.getAddress(), datagramPacket.getPort(), 50));
-		} else {
+			// guaranteed to be unique ID 128-bit value
+			//UUID id = UUID.randomUUID();
+			clients.add(new ServerClient(message.substring(3), datagramPacket.getAddress(), datagramPacket.getPort(), UniqueIdentifier.getIdentifier()));
+		} else if(message.startsWith("/m/")) {
+			sendToAll(message);
+		}
+		else {
 			System.out.println(message);
 		}
-
+	}
+	
+	private void sendToAll(String message) {
+		for(ServerClient client : clients) {
+			send(message.getBytes(), client.address, client.port);
+		}
+	}
+	
+	// Creates a new thread to send byte data to specified inet address and port in packet
+	public void send(final byte[] data, final InetAddress address, final int port) {
+		send = new Thread("Send") {
+			public void run() {
+				DatagramPacket datagramPacket = new DatagramPacket(data, data.length, address, port);
+				try {
+					datagramSocket.send(datagramPacket);
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			}
+		};
+		send.start();
 	}
 
 }