+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Project is still under construction
 
+## [0.1.1] - 2017-02-01
+### Added
+- Unique Identifier class that recycles unused id/s
+- Server relays message to all clients connected
+- refactored Client into front and back end
+
 ## [0.1.1] - 2017-01-24
 ### Added
 - ServerClients to handle information about clients connected to a server